requirements.txt
aiohttp==3.8.6
weaviate-client==4.6.0
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
blessed==1.20.0
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.2
filelock==3.12.4
frozenlist==1.4.0
fsspec==2023.9.2
greenlet==3.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
huggingface-hub==0.17.3
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
inquirer==3.1.3
inquirerpy==0.3.4
Jinja2==3.1.2
litellm==1.37.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.26.1
openai==1.28.1
packaging==23.2
pandas==2.1.1
pfzy==0.3.4
priority==2.0.0
prompt-toolkit==3.0.39
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
readchar==4.0.5
regex==2023.10.3
requests==2.31.0
rich==13.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tenacity==8.2.3
tiktoken==0.4.0
tokenizers==0.14.1
tokentrim==0.1.10
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
wcwidth==0.2.8
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0

